+++ config/module.config.php
@@ -1,10 +1,10 @@
 <?php
 
-use ZF\OAuth2\Doctrine\MutateTableNames\Container\MutateTableNamesSubscriberFactory;
-use ZF\OAuth2\Doctrine\MutateTableNames\EventSubscriber\MutateTableNamesSubscriber;
+use LaminasApi\OAuth2\Doctrine\MutateTableNames\Container\MutateTableNamesSubscriberFactory;
+use LaminasApi\OAuth2\Doctrine\MutateTableNames\EventSubscriber\MutateTableNamesSubscriber;
 
 return array(
-    'zf-oauth2-doctrine' => [
+    'apiskeletons-oauth2-doctrine' => [
         'mutatetablenames' => [
             'default' => []
         ],